address comments part 6: create engine class
